Broke out SubmissionBottomSheetActions.java from PopulateSubmissionViewHolder.java
